+++ CHANGES.md
@@ -2,3 +2,4 @@
 - 调整（删除）了一些选项，reformat 了一些代码
 - 所有 problem 均重命名为 challenge
 - 使用 gocat 替代 socat（使用 <https://github.com/taoky/gocat> 构建）
+- Rootless docker 支持（参见 example-bareflask）